requirements/base.txt
Django==2.2.6
django-extensions==2.2.5
pytz==2018.9
django-configurations==2.1
djangorestframework==3.9.1
Markdown==3.0.1
django-filter==2.1.0
gunicorn==19.9.0
django-storages==1.7.1
boto3==1.9.93
dj-database-url==0.5.0
django-model-utils==3.1.2
django_unique_upload==0.2.1
